Moved Elu operation from Fused to regular ones because the decomposition works extremely slowly.
